# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API non officielle, auto-hébergeable, qui transforme des pages publiques Manga-News en JSON.\n\nPoints importants :\n- pas de préfixe `/v1` ;\n- pas de routes admin publiques ;\n- authentification optionnelle via `Authorization: Bearer <API_TOKEN>` si `API_TOKEN` est défini ;\n- `ETag` / `If-None-Match` disponibles sur les réponses enveloppées ;\n- `X-Request-Id` renvoyé sur toutes les réponses pour corréler les logs ;\n- les réponses de recherche et de résolution peuvent être enrichies avec `title_vo`, `translated_title`, et, quand l'information existe, les compteurs `vf` / `vo` issus de la fiche série parente ;\n- les réponses volume restent légères par défaut et n'hydratent `vf` / `vo` que si `include_parent_editions=true` ou si la configuration serveur l'active explicitement ;\n- une route technique `/health/runtime` expose les métriques agrégées, les derniers événements de perf et l'état du cache local.\n\nFlux conseillé :\n1. utiliser `/search` ou `/search/resolve` pour obtenir un slug ou un couple `series_slug` / `volume_slug` ;\n2. consommer ensuite `/series/{slug}` ou `/volume/{series_slug}/{volume_slug}` ;\n3. réutiliser `ETag` et `If-None-Match` pour limiter les téléchargements inutiles.",
    "title": "Manga News Private API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/news/global": {
      "get": {
        "description": "Retourne le flux RSS global Manga-News converti en JSON.",
        "operationId": "get_global_news_news_global_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get global Manga-News RSS items",
        "tags": [
          "News"
        ]
      }
    },
    "/news/series/{slug}": {
      "get": {
        "description": "Récupère les actualités d’une série.",
        "operationId": "get_series_news_news_series__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get news for a series",
        "tags": [
          "News"
        ]
      }
    },
    "/news/volume/by-url": {
      "get": {
        "description": "Même comportement que `/news/volume/{series_slug}/{volume_slug}`, mais en partant d’une URL directe Manga-News.",
        "operationId": "get_volume_news_by_url_news_volume_by_url_get",
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "title": "Url",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get volume news from a direct URL",
        "tags": [
          "News"
        ]
      }
    },
    "/news/volume/{series_slug}/{volume_slug}": {
      "get": {
        "description": "Récupère les actualités d’un volume à partir de son couple `series_slug` / `volume_slug`.",
        "operationId": "get_volume_news_news_volume__series_slug___volume_slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "series_slug",
            "required": true,
            "schema": {
              "title": "Series Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "volume_slug",
            "required": true,
            "schema": {
              "title": "Volume Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get news for a volume",
        "tags": [
          "News"
        ]
      }
    },
...
    "/series/by-url": {
      "get": {
        "description": "Même comportement que `/series/{slug}`, mais en partant d’une URL directe Manga-News.",
        "operationId": "get_series_by_url_series_by_url_get",
        "parameters": [
          {
            "description": "URL complète de la fiche série Manga-News.",
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "description": "URL complète de la fiche série Manga-News.",
              "title": "Url",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "blocks",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Blocks"
            }
          },
          {
            "in": "query",
            "name": "fields",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fields"
            }
          },
          {
            "in": "query",
            "name": "include_raw_sections",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Raw Sections",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a series payload from a direct Manga-News URL",
        "tags": [
          "Series"
        ]
      }
    },
    "/series/by-url/editions": {
      "get": {
        "description": "Même comportement que `/series/{slug}/editions`, mais en partant d’une URL directe Manga-News.",
        "operationId": "get_series_editions_by_url_series_by_url_editions_get",
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "title": "Url",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "edition",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "all",
                "vf",
                "vo"
              ],
              "title": "Edition",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesEditionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List series editions from a direct URL",
        "tags": [
          "Series"
        ]
      }
    },
    "/series/by-url/related": {
      "get": {
        "description": "Même comportement que `/series/{slug}/related`, mais en partant d’une URL directe Manga-News.",
        "operationId": "get_series_related_by_url_series_by_url_related_get",
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "title": "Url",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesRelatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get related links from a direct series URL",
        "tags": [
          "Series"
        ]
      }
    },
...
    "/series/{slug}/editions": {
      "get": {
        "description": "Récupère les pages Éditions VF et/ou VO d’une série.",
        "operationId": "get_series_editions_series__slug__editions_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "description": "`all`, `vf` ou `vo`.",
            "in": "query",
            "name": "edition",
            "required": false,
            "schema": {
              "default": "all",
              "description": "`all`, `vf` ou `vo`.",
              "enum": [
                "all",
                "vf",
                "vo"
              ],
              "title": "Edition",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesEditionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List series editions",
        "tags": [
          "Series"
        ]
      }
    },
    "/series/{slug}/related": {
      "get": {
        "description": "Renvoie les liens en relation extraits depuis la fiche série.",
        "operationId": "get_series_related_series__slug__related_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesRelatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get related links for a series",
        "tags": [
          "Series"
        ]
      }
    },
...
  },
...
}